Checkpoint in the development:
Tag name: Osiris.

To implement:

1) early-stop technology (Not registred) (5-10 h)
2) paired-end reads deeds (5-10 h)
3) hypergraph path fusions (5-10 h)

Fix of the day:

The heuristic was wrong.

Roadmap/plans:

The code should  be completed by next week?

Although 100 % parallel, Ray code named Osiris is very similar to OpenAssembler version <submitted To Journal of Computational Biology>.
Adding a so-called merger would enable good assemblies with no duplications.

However, the hypergraph paths fusions approach is superior because it leads to the usage of our yet-to-be-implemented early-stop technology, which cut clear in the computational time (it allows the program to avoid parts of the graph that are strongly assembled already.)

Sébastien Boisvert.