Modified the code that writes scaffolds in FASTA format.
Instead of sending an MPI message to an MPI rank to tell it to open
the FASTA file and to append to it its contig, the MASTER rank is the only
one that actually write bits to the file. Other MPI ranks are just asked
for their important bits.

I initially thought that writting scaffold sequences was slow because
of the numerous fprintf (one for each nucleotide), but it seems that
it is not the case. I believe it was slow because of the overhead of
all the fopen/fprintf/fclose that were generated on numerous MPI ranks on the
same very file which is available through a net filesystem or Lustre.

I guess this will fix the problem.