This commit finally fixes a very critical bug that happens once in a while. This bug is critical and v1.7 is affected. It causes strange behaviors usually leading to segmentation faults or Ray errors.
In fact, the bug appears randomly and only on certain versions of the tool chain
and only on particular super-computers (see the diff).

The bug happened because forceFlush() can not be called if there
are active workers ready to do some work.

Here, forceFlush() was called without checking if there were any
active workers.

Checking for active workers before calling forceFlush fixes
the problem.

This is already fixed upstream
in RayPlatform and all use cases where the VirtualProcessor is utilised
should be ported to the TaskCreator interface to weed out these bugs.