Searcher: fixed a race condition
Pier-Luc Plante and Sébastien Boisvert worked on this bug that
caused Ray to hang in a specific condition triggered by a race
condition.

The bug occured within one single iteration of the RayPlatform
virtual machine. A message of type
RAY_MPI_TAG_SEARCH_ELEMENTS_REPLY was first received. But it was
not consumed during the same iteration because all the files in
a directory were synced, therefore it was forever lost.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>