SeedingData: seeds can not contain k-mers with too low coverage
The distributed de Bruijn graph may contain k-mers with a coverage
of 1 for various reasons, including false positives produced by
the Bloom filter. For this reason, k-mers with a coverage that is
below the minimum required by the policy should be discarded.
The plugin SeedExtender does that already, but it seems that
the plugin SeedingData should do it too.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>