MessageProcessor: k-mer data messages should never be discarded
In cases where the lexicographically-lower k-mer of a pair of
k-mers occur after the other in the sequence data and where the
collective k-mer coverage of the pair is the minimum accepted in
by the distributed storage engine, race conditions can
occur. These race conditions will produce an invalid degree
distribution where, for example, the number of vertices (k-mers)
with 1 parent and 2 children is not stricly the same than the
vertices with exactly 2 parents and 1 child while it should be the
case. Another consequence of this race condition is that
the graph will contain (u,v) in the directed edges, but not
(r(v),r(u)) where r(.) is the reverse complement of its operand.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>