KmerAcademyBuilder: only send the forward k-mer, not the lower
This patch has no impact on the result generated by the assembly.
Basically, each k-mer has a reverse-complement and sending both
will increase by one the coverage of both. But in the
implementation, a forward k-mer and its reverse-complement are
stored together to save memory. So, for any forward k-mer occurring
in a read (which can come from the forward strand or from the
reverse-complement strand) sending only the forward k-mer (as is)
is totally equivalent to sending the lowest k-mer between the
forward and the reverse-complement. But the code is clearer if it
sends simply just the forward instead of the lowest.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>