MessageProcessor: don't discard k-mers while receiving messages
This patch is a fix for the regression introduced by the
recent patch dc65e18327870ad5f57bb75d7b4760e0141e6493.
The plugin KmerAcademyBuilder was modified to be clearer and to fix a bug.
The change was to send any forward k-mer, not just the lower of a pair.
However, the current plugin sends messages to another one, which was still
implementing the old behavior. The handlers
call_RAY_MPI_TAG_VERTICES_DATA and call_RAY_MPI_TAG_KMER_ACADEMY_DATA
were modified accordingly.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>