VerticesExtractor: store twin edges in a single source
In Ray, the k-mers are stored in pairs. These twin k-mers are
reverse-complement. A k-mer can not be its own twin because Ray
disallows odd k-mer lengths. This patch remove the duplicated
data source for twin edges. Now, the k-mer and its reverse-
complement are stored together, and its edges and the edges
of its reverse-complement are also stored together.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>