NetworkTest: added average round trip latency
The average round trip latency was added to the network test. Now,
there are the mode round trip latency and the average round trip
latency in the results. Passing a message of the type
RAY_MPI_TAG_TEST_NETWORK_MESSAGE and getting a reply of the type
RAY_MPI_TAG_TEST_NETWORK_MESSAGE_REPLY is faster between processes
on the same machine. Usually, this means that the mode latency is
a little bit lower than the average latency, globally.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>