Searcher: find or create a virtual color from physical colors
The code for finding or creating a virtual color from physical
colors was added. The new code is only used for storing summary
information. The code that handles virtual coloring for kmers in
the graph is much more efficient as it is O(1) for most operations.

In particular, it is faster to obtain a hash value from a virtual
color and a physical color than with just an array of physical
colors.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>