Searcher: added physical color in SequenceAbundances.xml
To make sense of the file
orion/BiologicalAbundances/_Coloring/GraphBrowsing.xml, the file
orion/BiologicalAbundances/DatabaseTest-strept/SequenceAbundances.xml
(or any other namespace directory) is required.

GraphBrowsing.xml is usefull -- it contains information about
coloring. For instance, you can obtain the number of references that
have colors 0 and 1 in namespace 0. For the biologist, this is
useful because it gives the number of k-mers shared by reference
sequences, for instance.

Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>