fix the gossip algorithm for seed merging
I tested the algorithm discussed with François Laviolette
and it did not work on ERS006494 (258 326 980 sequences).
Indeed, it was slow and generated too many messages
on the fabric.

I reanalyzed the designs that Élénie Godzaridis provided for
the gossip algorithm. The 'linked actors' must be generated
using the endpoints of gossip assets. The initial
implementation used a set of linked actors for each rank,
regardless of these endpoints (which are paths).

Link: http://en.wikipedia.org/wiki/Gossip_protocol
Signed-off-by: Sébastien Boisvert <sebastien.boisvert.3@ulaval.ca>